+++ dmn-pmml-springboot-example/src/test/java/org/kie/kogito/dmn/pmml/springboot/example/CommonTestUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.dmn.pmml.kogito.springboot.example;
+package org.kie.kogito.dmn.pmml.springboot.example;
 
 import java.util.Map;
