+++ dmn-pmml-springboot-example/src/test/java/org/kie/kogito/dmn/pmml/springboot/example/MiningModelTest.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.dmn.pmml.kogito.springboot.example;
+package org.kie.kogito.dmn.pmml.springboot.example;
 
-import java.util.HashMap;
+import java.util.Collections;
 import java.util.Map;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -26,17 +26,17 @@ import org.springframework.test.annotation.DirtiesContext;
 
 import io.restassured.RestAssured;
 
-import static org.kie.dmn.pmml.kogito.springboot.example.CommonTestUtils.testDescriptive;
-import static org.kie.dmn.pmml.kogito.springboot.example.CommonTestUtils.testDescriptiveWrongData;
-import static org.kie.dmn.pmml.kogito.springboot.example.CommonTestUtils.testResult;
-import static org.kie.dmn.pmml.kogito.springboot.example.CommonTestUtils.testResultWrongData;
+import static org.kie.kogito.dmn.pmml.springboot.example.CommonTestUtils.testDescriptive;
+import static org.kie.kogito.dmn.pmml.springboot.example.CommonTestUtils.testDescriptiveWrongData;
+import static org.kie.kogito.dmn.pmml.springboot.example.CommonTestUtils.testResult;
+import static org.kie.kogito.dmn.pmml.springboot.example.CommonTestUtils.testResultWrongData;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = KogitoSpringbootApplication.class)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
-public class DecisionTreeTest {
+public class MiningModelTest {
 
-    private static final String BASE_PATH = "/DecisionTree";
-    private static final String TARGET = "decision";
+    private static final String BASE_PATH = "/MiningModelSum";
+    private static final String TARGET = "result";
 
     @LocalServerPort
     private int port;
@@ -47,29 +47,27 @@ public class DecisionTreeTest {
     }
 
     @Test
-    void testEvaluateDecisionTreeResult() {
-        String inputData = "{\"temperature\":30.0, \"humidity\":10.0}";
-        testResult(inputData, BASE_PATH, TARGET, "sunglasses");
+    void testEvaluateMiningModelResult() {
+        String inputData = "{\"input1\":200.0, \"input2\":-1.0, \"input3\":2.0}";
+        testResult(inputData, BASE_PATH, TARGET, -299.0f);
     }
 
     @Test
-    void testEvaluateDecisionTreeResultWrongData() {
-        String inputData = "{\"temperature\":\"b\", \"humidity\":10.0}";
+    void testEvaluateMiningModelResultWrongData() {
+        String inputData = "{\"input1\":wrong-data, \"input2\":-1.0, \"input3\":2.0}";
         testResultWrongData(inputData, BASE_PATH);
     }
 
     @Test
-    void testEvaluateDecisionTreeDescriptive() {
-        String inputData = "{\"temperature\":30.0, \"humidity\":10.0}";
-        final Map<String, Object> expectedResultMap = new HashMap<>();
-        expectedResultMap.put(TARGET, "sunglasses");
-        expectedResultMap.put("weatherdecision", "sunglasses");
+    void testEvaluateMiningModelResultDescriptive() {
+        String inputData = "{\"input1\":200.0, \"input2\":-1.0, \"input3\":2.0}";
+        final Map<String, Object> expectedResultMap = Collections.singletonMap(TARGET, -299.0f);
         testDescriptive(inputData, BASE_PATH, TARGET, expectedResultMap);
     }
 
     @Test
-    void testEvaluateDecisionTreeDecriptiveWrongData() {
-        String inputData = "{\"temperature\":\"b\", \"humidity\":10.0}";
+    void testEvaluateMiningModelResultDescriptiveWrongData() {
+        String inputData = "{\"input1\":wrong-data, \"input2\":-1.0, \"input3\":2.0}";
         testDescriptiveWrongData(inputData, BASE_PATH);
     }
 }
